+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 * Diff-view for version history
+* Set form fields readonly instead of disabled
 
 
 ## v2024.3 - 2024-01-09